+++ app-backend/common/src/main/scala/datamodel/ProjectLayer.scala
@@ -1,23 +1,29 @@
 package com.rasterfoundry.common.datamodel
 
+import com.rasterfoundry.common.JsonCodecs
 import com.typesafe.scalalogging.LazyLogging
 import geotrellis.vector.{Geometry, Projected}
+import io.circe.{Decoder, Encoder}
 import io.circe.generic.JsonCodec
+import io.circe.generic.semiauto._
 
 import java.sql.Timestamp
 import java.util.UUID
 
 @JsonCodec
-final case class ProjectLayer(id: UUID,
-                              createdAt: Timestamp,
-                              modifiedAt: Timestamp,
-                              name: String,
-                              projectId: Option[UUID],
-                              colorGroupHex: String,
-                              smartLayerId: Option[UUID],
-                              rangeStart: Option[Timestamp],
-                              rangeEnd: Option[Timestamp],
-                              geometry: Option[Projected[Geometry]])
+final case class ProjectLayer(
+    id: UUID,
+    createdAt: Timestamp,
+    modifiedAt: Timestamp,
+    name: String,
+    projectId: Option[UUID],
+    colorGroupHex: String,
+    smartLayerId: Option[UUID],
+    rangeStart: Option[Timestamp],
+    rangeEnd: Option[Timestamp],
+    geometry: Option[Projected[Geometry]],
+    isSingleBand: Boolean,
+    singleBandOptions: Option[SingleBandOptions.Params])
     extends GeoJSONSerializable[ProjectLayer.GeoJSON] {
   def toGeoJSONFeature: ProjectLayer.GeoJSON = ProjectLayer.GeoJSON(
     this.id,
@@ -30,32 +36,37 @@ final case class ProjectLayer(id: UUID,
       this.colorGroupHex,
       this.smartLayerId,
       this.rangeStart,
-      this.rangeEnd
+      this.rangeEnd,
+      this.isSingleBand,
+      this.singleBandOptions
     )
   )
 }
 
 @JsonCodec
-final case class ProjectLayerProperties(projectId: Option[UUID],
-                                        createdAt: Timestamp,
-                                        modifiedAt: Timestamp,
-                                        name: String,
-                                        colorGroupHex: String,
-                                        smartLayerId: Option[UUID],
-                                        rangeStart: Option[Timestamp],
-                                        rangeEnd: Option[Timestamp])
+final case class ProjectLayerProperties(
+    projectId: Option[UUID],
+    createdAt: Timestamp,
+    modifiedAt: Timestamp,
+    name: String,
+    colorGroupHex: String,
+    smartLayerId: Option[UUID],
+    rangeStart: Option[Timestamp],
+    rangeEnd: Option[Timestamp],
+    isSingleBand: Boolean,
+    singleBandOptions: Option[SingleBandOptions.Params])
 
-object ProjectLayer extends LazyLogging {
-  def create = Create.apply _
+object ProjectLayer extends LazyLogging with JsonCodecs {
 
-  @JsonCodec
   final case class Create(name: String,
                           projectId: Option[UUID],
                           colorGroupHex: String,
                           smartLayerId: Option[UUID],
                           rangeStart: Option[Timestamp],
                           rangeEnd: Option[Timestamp],
-                          geometry: Option[Projected[Geometry]]) {
+                          geometry: Option[Projected[Geometry]],
+                          isSingleBand: Boolean,
+                          singleBandOptions: Option[SingleBandOptions.Params]) {
     def toProjectLayer: ProjectLayer = {
       val now = new Timestamp(new java.util.Date().getTime)
       ProjectLayer(
@@ -68,10 +79,55 @@ object ProjectLayer extends LazyLogging {
         this.smartLayerId,
         this.rangeStart,
         this.rangeEnd,
-        this.geometry
+        this.geometry,
+        this.isSingleBand,
+        this.singleBandOptions
       )
     }
   }
+
+  object Create {
+    implicit val createEncoder: Encoder[Create] = deriveEncoder
+    implicit val createDecoder: Decoder[Create] = Decoder.forProduct7(
+      "name",
+      "projectId",
+      "colorGroupHex",
+      "smartLayerId",
+      "rangeStart",
+      "rangeEnd",
+      "geometry"
+    )(
+      (name: String,
+       projectId: Option[UUID],
+       colorGroupHex: String,
+       smartLayerId: Option[UUID],
+       rangeStart: Option[Timestamp],
+       rangeEnd: Option[Timestamp],
+       geometry: Option[Projected[Geometry]]) => {
+        Create(
+          name,
+          projectId,
+          colorGroupHex,
+          smartLayerId,
+          rangeStart,
+          rangeEnd,
+          geometry,
+          false,
+          None
+        )
+      }) or Decoder.forProduct9(
+      "name",
+      "projectId",
+      "colorGroupHex",
+      "smartLayerId",
+      "rangeStart",
+      "rangeEnd",
+      "geometry",
+      "isSingleBand",
+      "singleBandOptions"
+    )(Create.apply _)
+  }
+
   final case class GeoJSON(id: UUID,
                            geometry: Option[Projected[Geometry]],
                            properties: ProjectLayerProperties,
@@ -88,7 +144,9 @@ object ProjectLayer extends LazyLogging {
         properties.smartLayerId,
         properties.rangeStart,
         properties.rangeEnd,
-        geometry
+        geometry,
+        properties.isSingleBand,
+        properties.singleBandOptions
       )
     }
   }