+++ app-backend/db/src/main/scala/ProjectLayerDao.scala
@@ -15,7 +15,7 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
   val tableName = "project_layers"
 
   val selectF: Fragment =
-    fr"SELECT id, created_at, modified_at, name, project_id, color_group_hex, smart_layer_id, range_start, range_end, geometry from" ++ tableF
+    fr"SELECT id, created_at, modified_at, name, project_id, color_group_hex, smart_layer_id, range_start, range_end, geometry, is_single_band, single_band_options from" ++ tableF
 
   def unsafeGetProjectLayerById(
       projectLayerId: UUID): ConnectionIO[ProjectLayer] = {
@@ -35,11 +35,11 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
   ): ConnectionIO[ProjectLayer] = {
     (fr"INSERT INTO" ++ tableF ++ fr"""
     (id, created_at, modified_at, name, project_id, color_group_hex,
-    smart_layer_id, range_start, range_end, geometry)
+    smart_layer_id, range_start, range_end, geometry, is_single_band, single_band_options)
     VALUES
       (${pl.id}, ${pl.createdAt}, ${pl.modifiedAt}, ${pl.name}, ${pl.projectId},
       ${pl.colorGroupHex}, ${pl.smartLayerId}, ${pl.rangeStart}, ${pl.rangeEnd},
-      ${pl.geometry})
+      ${pl.geometry}, ${pl.isSingleBand}, ${pl.singleBandOptions})
     """).update.withUniqueGeneratedKeys[ProjectLayer](
       "id",
       "created_at",
@@ -50,7 +50,9 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
       "smart_layer_id",
       "range_start",
       "range_end",
-      "geometry"
+      "geometry",
+      "is_single_band",
+      "single_band_options"
     )
   }
 
@@ -62,7 +64,9 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
       name = ${projectLayer.name},
       color_group_hex = ${projectLayer.colorGroupHex},
       geometry = ${projectLayer.geometry},
-      project_id = ${projectLayer.projectId}
+      project_id = ${projectLayer.projectId},
+      is_single_band = ${projectLayer.isSingleBand},
+      single_band_options = ${projectLayer.singleBandOptions}
     """ ++ Fragments.whereAndOpt(Some(idFilter))).update
     query
   }